refactor: 마이크로 프론트엔드 적용
- turbo 사용
- 하위 패키지 node_modules 포함되지 않도록 처리
- 불필요한 환경 변수들 제거
