+++ @types/env.d.ts
@@ -1,10 +1,6 @@
 namespace NodeJS {
   interface ProcessEnv {
-    NEXT_NOTION_API_USER_ID: string;
     NEXT_NOTION_API_AUTH_TOKEN: string;
-    NEXT_NOTION_API_TOKEN_V2: string;
-    NEXT_NOTION_API_BASE_URL: string;
-    NEXT_NOTION_API_DATABASE_ID: string;
     NEXT_NOTION_API_DATA_SOURCE_ID: string;
   }
 }
