+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@einere/common-utils": "^1.0.11",
     "@giscus/react": "^3.1.0",
+    "@homepage/notion-renderer": "workspace:*",
     "@notionhq/client": "^5.0.0",
     "@tanstack/react-query": "^5.83.0",
     "@tanstack/react-query-devtools": "^5.83.0",
@@ -38,6 +39,7 @@
     "prettier": "3.5.3",
     "prettier-plugin-tailwindcss": "^0.6.11",
     "tailwindcss": "^4",
+    "turbo": "^2.0.0",
     "typescript": "^5"
   },
   "packageManager": "pnpm@7.33.6"
