--- src/app/components/NotionRenderer/index.ts
@@ -1,7 +1,0 @@
-export { NotionRenderer } from "./NotionRenderer";
-export type {
-  NotionBlock,
-  NotionBlockList,
-  NotionRendererProps,
-} from "./types";
-export { RichText } from "./components/RichText";
